# [CHIA-1829] don't drop outgoing response messages 