@arvidn 2024-12-06
the missing test coverage is because we (currently) don't have any message type that's a "tx" type message that's unlimited. Not supporting that now could cause confusion in the future however.