@arvidn 2024-12-06
I'm having second thoughts about this solution. using `None` to indicate unlimited also removes the ability to limit the max size of a single message.